Create Custom Upstream - Add note stating this feature is available for certain partner tiers
The notice is to make developers aware that adding to custom upstreams is only available to partners - not Ally. I had a dev go through the trouble of creating a custom upstream only to learn later that it is only available to partners. He would have appreciated that detail in this doc before spending time creating a custom upstream.

Add link to create agency profile, use HTML instead of MD within note div