Update domains.md
This is when FF and Safari trying to access a not properly setup bare domain or has not propagated yet but www.example.com is already accessible. FF and Safari will prepend `www.` so it will automatically work using www.example.com instead. While chrome will give the unresolved error or go to google search and use the typed word as the search keyword.