fix logic for multidev
Logic is backwards. Let's be positive. We should define TEST and LIVE, then catch the rest: DEV, and multidev. Not only is this clearer to read, but as it stood we would be implementing production cache settings on multidev environments. Perhaps it is desirable to set multidevs to mirror production, like we do TEST, but that really depends on workflow.